Add npm package size check utility
This commit introduces a new Go utility that checks the unzipped size of all packages in an npm organization. The utility fetches the packages, retrieves their sizes, and output the results to a CSV file and the console in the form of a table. A new README file is also added to help users with the setup and usage of the utility.